+++ src/main/kotlin/joshxviii/plantz/entity/plant/Cactus.kt
@@ -30,7 +30,7 @@ class Cactus(type: EntityType<out Plant>, level: Level) : Plant(type, level) {
             random: RandomSource
         ): Boolean {
             val blockBelow = level.getBlockState(pos.below())
-            return checkValidSpawn(level, pos, spawnReason)
+            return checkValidSpawn(level, pos, spawnReason, random)
                     && (blockBelow.`is`(PLANTABLE) || blockBelow.`is`(BlockTags.SAND) || blockBelow.`is`(Blocks.SOUL_SAND))
         }
     }