+++ src/main/kotlin/joshxviii/plantz/entity/plant/IcePeaShooter.kt
@@ -29,7 +29,7 @@ class IcePeaShooter(type: EntityType<out Plant>, level: Level) : Plant(type, lev
             random: RandomSource
         ): Boolean {
             val blockBelow = level.getBlockState(pos.below())
-            return checkValidSpawn(level, pos, spawnReason)
+            return checkValidSpawn(level, pos, spawnReason, random)
                     && (blockBelow.`is`(PLANTABLE) || blockBelow.`is`(BlockTags.SNOW))
         }
     }