+++ src/main/kotlin/joshxviii/plantz/entity/plant/SeaShroom.kt
@@ -31,7 +31,7 @@ class SeaShroom(type: EntityType<out Plant>, level: Level) : Plant(type, level) 
             val isRaining = level.level.isRaining
             val rainBonus = if (isRaining) 2.25f else 1f
 
-            return checkWaterSpawn(level, pos, spawnReason)
+            return checkWaterSpawn(level, pos, spawnReason, random)
                     && random.nextFloat() < (0.25 * rainBonus)
         }
     }